requirements.txt
pushbullet.py==0.9.0
python-magic==0.4.10
requests==2.8.1
six==1.10.0
websocket-client==0.34.0
wheel==0.24.0
